backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
SQLAlchemy==2.0.25
PyJWT==2.8.0
Werkzeug==3.0.1
numpy==1.26.4
joblib==1.4.2
scikit-learn==1.4.2
pandas==2.2.2
matplotlib==3.8.4
fpdf==1.7.2
psycopg2-binary==2.9.9

# Phase 2 additions
matplotlib==3.8.4
pyjwt==2.8.0
werkzeug>=3.0.1

